requirements.txt
requests>=2.31.0,<3
speedtest-cli>=2.1.3,<3
matplotlib>=3.7,<4
